users/pyproject.toml
[project]
name = "users"
version = "0.1.0"
description = "Flask app reading users from YAML"
authors = [
    {name = "Joshua Darovitz",email = "joshdarovitz@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
